+++ src/ray_tracer.cpp
@@ -1,7 +1,7 @@
 #include <math.h>
 #include "ray_tracer.h"
 
-bool rayTracer::equalWithError(const double &lhs, const double &rhs) {
+bool rayTracer::equalWithError(const float &lhs, const float &rhs) {
     if (abs(lhs - rhs) < EPSILON)
         return true;
     return false;
